Change how dsn is done
Before I tried to just have it injected from an env var at build time,
but this didn't work right away. It probably could work eventually but
it doesn't really buy much since anyone with the binary can get the DNS
out if they try. So instead just put the DSN in the code. This has some
very light obfuscation just to thwart automated github scanners looking
for secrets, but that's it.