Make setting actions a bit easier
Now that most of the actions in the cli parser are proper Actions, it
felt very repetitive to always have to call new and pass in the client.
So this adds a very simple macro that does that and assigns it to the
`action` local variable. You can still assign the specific action to
another local variable to make sure everything is typed correctly.