Add cluster role methods to client.
Adds cluster role methods to `CB::Client`. As well, updates calls to get
the default role to use new method that allows specifying the name of
the cluster role to retrieve. The URI on the `role` record was updated
to be nilable, as it is possible for the API to return a null value for
URI when a role is deleted.