Make `host` configuration global to the application.
Previously, the host configuration was being read from ENV in multiple
locations and then passed to different classes and methods.  These
changes move that initialization to happen early on in startup and sets
a constant for it at the top level of the CB module. Therefore it is
only ever read/set once per invocation of the application.