Add some error handling to `cb logs`.
It appears that no exception handling was being done by this action.
Therefore it would output stacktraces to the user. While this doesn't
necessarily make the cause of the error easier to determine for the
user. It is at least a better look.  It would be great to tie in some
reporting on some of these, but that's for a potential future update.