Refactor `login` to an `Action`.
This refactor required a bit more than just moving the function from
`Program` to an action. In order to do this, we first need to
distinguish between actions that 'require' a client and those that
don't.

For example, `TeamList` is an action that requires access to the API and
therefore needs an client configured. However, `Login` simply collects
and persists the login credentials used to obtain a token. This action
does not require any API access.