Add `CB::Role` type.
Here we're pulling up common role and role name operations and stuctures
into one place.  We also add a new `role_setter` that also allows us to
pull up roll name validation up to when the value is set. This is useful
as it we can DRY up some of the role related code as well as simplify
some of the test cases.