Update `Spectator` to `0.11.3`.
This is a pretty large changeset due in part to the breaking changes
that were introduced by this update.

We also take the opportunity to go ahead and transition the remaining
tests to bring them in line.

There were some interesting challenges. Mostly around how the create
cluster action behaves. Specifically, it's the related client calls that
caused a bit of a head ache. For some reason, Spectator doesn't like how
we were doing it. And when stubbing the `create_cluster` method, things
'worked' (verified by tracing the call), but for some reason it would
cause a crash elsewhere that was quite annoying. I believe it's possible
that there is an issue in Spectator around this. However, I was unable
to pin point it. However, I'm not terribly concerned with it as making
the transition on these client methods has been on the TODO list for
some time. So, we're also taking that opportunity with these changes.