Add `Format` enum.
Here we're adding a more generalized format enum for output formats.

Previously, we had started to define this at the action level as it was
more about experimenting with it. But now it seems like something we
should likely just pull up to the top level so that all actions can
benefit from it at some point.