Update `TeamInfo` action to use `Format::List`.
This change does not introduce new functionality. It's just a refactor
of the TeamInfo action to incorporate the use of the `Format::List`
enum.

As well, we introduce the use of the using the `TableBuilder` to create
a list output. The output is the same. But the method is different in
that by using the table builder we don't have to worry about layout of
the values.

As well, this sets us up to start offering other formats to the rest of
the team actions. However, it is not possible to specify the format just
yet. That will be included in later changes.