Allow API key secret-only login and usage
API keys have both an application ID and application secret, and
previously both of them were required to use one. At some point we
realized this hadn't really been entirely thought through and wasn't
really necessary since API key secrets already have plenty of entropy by
themselves, and as of mid-June last year we started to allow secret-only
authentication in the API. Newer secrets were prefixed with `cbkey_` so
that we could easily distinguish an API key secret from other types of
secrets that might be used for authentication (like an access token
secret).

Here, support secret-only authentication in cb. This makes `cb login` a
bit faster to use since you only need to copy in one value.

The downside to the approach I've built here is that we'd have to force
anyone with an older API key to rotate it to continue using a new
version of cb. I figure this might be okay though since we've been
issuing newer style API keys for seven months now, and a lot of people
will already have rotated theirs organically.

The other alternative is that we could continue to try and read an
application ID from `creds` but just stop storing one. When reading one,
allow ID to be empty on newer API keys, but use ID if an old one is
detected, and only error in the case there's no ID and the API key is
old. This would have improved backward compatibility, but make the code
a bit cruftier and more complex. I opted for the simpler alternative,
but don't mind implementing it this way if it's preferable.