Implement command for `cb open` to open a Dashboard session
Here, we implement a new command: `cb open`. It uses the API key bundled
into cb to create a session via API, then pops a web browser using a
system command like `open` on Mac or `xdg-open` on Linux to log the user
into the session via use of a one-time token (see [1] for background).

I'm not fully decided on the longer term survival of this feature (at
least with its current implementation) because although secure, it does
have does have to make a concession that's suboptimal for security.
Because commands like `open` can only support a `GET` request, one-time
tokens are sent via query parameter instead of body, which means they
could conceivably land in a log somewhere, even if the request itself is
operating over a secure channel.

We do take steps to keep one-time tokens secure -- they're named such
because they're always burned the first time the API ever sees one,
regardless of whether the request succeeds or not. So by the time it
ever leaks into a log, it should already be useless. One-time tokens are
also only usable within one minute of creation, so it's not possible to
not use on and have it linger around as a possible backdoor.

Still, it'd probably be overall better if more of this functionality was
pushed into the browser, but to make that happen, the Frontend would
have to be fully aware of login functionality, which is a much larger
project given its pace of change. So the idea here is to get something
in that works, and then come up with a more optimal fully-formed solution
later on.

In order to not make people too dependent on `cb open` in case we want
to change something, it's hidden from the `cb help` list so that it'll
generally only be used by those who already know about it.

[1] https://github.com/CrunchyData/priv-all-cloud-owl/pull/2252