Add `CB::Types::Tree` data type.
Here we're adding a new data structure type that represents a simple
tree.

The impetus for adding support for this type is that sometimes it is
necessary to organize output data in a tree format. By adding this data
structure it makes it easier for us to traverse the the data in a way
that facilitates it's output in such a format.

We do not add the actual rendering of the output here as this is just
intended to introduce the data structure. How it is rendered will be
following in a later commit.