Add `CB::Tree::Renderer`.
Here we're adding a new renderer that will output data in tree format.

This renderer utilizes the `Tree` stucture that was introduce in a prior
commit.  The goal is output data in a tree format that represents a
'depth first' traversal of the tree data.

An example of what the output might look like for a simple tree:

```
root
├─ first
├─ second
└─ third
```

This renderer can also handle trees that have an arbitrary number of
levels.  For example:

```
root
├─ first
│  ├─ child 1
│  │  └─ grandchild 1
│  └─ child 2
├─ second
└─ third
```