Add `flatten` param to `get_clusters` client method.
Here we add a new parameters to the `get_clusters(teams)` method.  The
effect of this parameter is to flatten out all of the clusters and
replicas. The purpose of this is to expose the replicas at the top level
of the result so that in case there are multiple levels of replicas they
are readily available for things like `autocomplete`, etc.

To clarify the effect of this parameter, all replicas are moved up and
the cluster's replica array is set to nil.