Update Makefile
Updates how the `OPENSSL_LIB_PATH` and `LIBSSH2_LIB_PATH` are
determined.  Previously, these paths were determined using `$(shell brew
--prefix <lib>)/lib`. However, the other libs utilize `pkg-config` for
this purpose. This is a problem for environments that do not use `brew`.
As well, it's also an inconsistency that seems worth cleaning up.

Also, we append the existing `LIBRARY_PATH` to the `LIBRARY_PATH`
defined in the `spec` target when `darwin` is the `TARGET_OS`. This,
again, is a band-aid for environments that do not use `brew` and might
have `LIBRARY_PATH` defined, e.g. through `direnv`.