Add `--team` option to `cb list`.
Here were implementing the `--team` option in `cb list`.  The flag was
inadvertently added in #97, but the intent was always to follow up with
this feature. So, we go ahead and make it happen here.

Ultimately, the idea is to allow a user to specify an individual team to
filter the results. As opposed to always returning all clusters for all
teams which they have access to.