Remove some `.psqlrc` configurations for `cb psql`.
Here we're simply removing all but the prompt configuration that is
placed in the temporary `.psqlrc` file that the `cb psql` command relies
on.  The impetus for this is that some users have reported unexpected
behaviors based on what they already have defined in their own
`.psqlrc`.  So, in order to respect what they already have configured,
we remove any potential behavior configuration from our generated
configuration for a psql session.

However, as noted above, we do keep the prompt configuration as that
seems like a reasonable configuration to keep for informational purposes
and it does not have any unexpected affect on the clients behaviors.