Add `CB::Model::Cluster`.
Here we're refactoring `CB::Client::Cluster` to `CB::Model::Cluster`.
As well, as part of this refactor, we are merging it with
`CB::Client::ClusterDetails` as the separation is no longer necessary.

The motivation for this is the following reasons:

* Cleanining up the Client such that it is only concerned with making
  client calls.
* Decoupling the use of the API resource data models from the Client.
  Basically, better use of namespacing and ultimately better
  organization of the code.

We also breakout the `ClusterStatus` into it's own model and add a
separate endpoint point for it as the API has deprecreated these items
from the `/clusters` endpoint and moved them to `/clusters/:id/status`.
This helps bring the client more up-to-date with the API.