Refactor `cb maintenance info` output.
Here we're refactoring how the output for `cb maintenance info` is
produced.

Specifically, we're moving the output code out of the `Operation` model
and putting it into the action. As well, we make use of the TableBuilder
that we are using in other command outputs, using it similar to how we
use it with `cb team list`.

Also, as part of this we guarantee the ordering of the output elements.
Specifically, we ensure that operations are always first and the
`maintenance window` is always last.