Add flake for devShells
This flake adds dev shells for x86 linux and both x86/arm mac
development. Right now it's using a crystal flake I made in my own
github account, but we can move that to crunchy's nixpkgs once we get
that going.

The flake does not provide a build yet for cb, but that should be
possible in the future.