Use nix for checks and CI
This adds two things. First a single command `check` that runs
formatting checks, ameba checks, and the specs not only in parallel,
which is cool enough, but also each check knows what files it depends on
and wont run if those haven't changed. So if you run `check` and then
edit the readme, it wont run the specs again.

Second it uses these checks on CI and also caches the check builds
between pull requests. So also on CI we can have all the checks only run
if they need to. And all of this without needing to do ever more careful
github caching and filters in yaml and so on. Just a single job that
takes care of everything.