Add support for cross platform creation of forks and replicas.
Here we're updating `cb create --fork` and `cb create --replica` to
support cross platform creation. Previously, this was not possible as
the fork and replica client requests were always defaulting to the
target/primary cluster's platform, region and plan. Now, it is possible
to create a fork or replica in another platform in the same way it is
that a standalone cluster is created.

For instance, to create an AWS based fork in GCP:

```
$ cb create --fork <ID> --platform gcp --region uswest-1 --plan
standard-8
```

If `--name` is omitted, then the value will default to `Fork of <name>`
where `<name>` is the name of the source cluster.  The same is true for
replicas, expept that it will `Replica of <name>`.

This also allows for the creation of forks and replicas across regions
within the same platform as well. Though, it still requires the full
qualification of specifying the target `--platform` and `--plan`.
Future improvements on this might be possible to simply allow the user
to only specify the `--region`. However, for now it seems prudent to
simply keep this current requirement as it helps to ensure (through
completion) that the any region or plan specified is correct/available
for a specific platform.