Escape team name in `psql` prompt.
Recently, Bridge moved to treating personal teams as regular teams. As a
result of that change, personal teams were renamed to `<first name>'s
Team`, e.g. `Adam's Team`. Unfortunately, however, the `'` was not being
properly escaped in the `PROMPT1` defined in `cb psql` command. As a
quick fix to this, we're simply adding a function to escape all
single-quotes in the team name before setting the prompt.