Update `whoami` to fetch user information from API.
Previously, `whoami` was relying fetching information from a token. The
problem with this is that this information might not always be
up-to-date or accurate. As a token is cached locally and the a different
user might have 'logged in'. And since the 'login' process does not
update the current token (this happens once the token expires,
admittedly a poor design, but improvements are incoming), a user could
'login' as a different user but still be given the previous user info
from `whoami`. This is certainly a 'bug' and this is the just the first
step in fixing this issue.