Update `cb token`.
For the most part this is a non-functional refactoring of the `cb token`
command. Though, it does improve how it works and is intended to be used
used. For instance, previously it wasn't entirely obvious that an API
key was necessary for this command to succeed as the API key was the
only credential being stored on the system. As well, it was using
mechanisms that were intended to support obtaining a token for use with
other commands.  `cb` is now moving in the direction of eliminating the
API as the primary credential while still allowing it to be provided via
ENV (`CB_API_KEY`) or by adding it manually to the credentials store.

The only functional change is that we now support the `--format` option
on the command. Specifically, `--format=json` and `--format=header` are
now supported. However, the use of `--format` and `-H` cannot be
used at the same time. As this will create a conflict and the command
does not define which should take precedence at this time.