Refactor `cb open`.
Refactors the `cb open` command by wrapping the `open` and/or `xdg-open`
command in it's own static class.  The reason for this change is to
allow other commands to be able to leverage these system commands when
necessary without having to reimplement the logic around them.

We've set this class up with two class methods: `run` and `exec`. Both
of these methods reflect the same behavior as their analogs in
`Process`.

For reference:

* `Open.run` - executes the `open` command as a process and waits for it
  to complete.

* `Open.exec` - replaces with the current process with a new one that
  runs the `open` command.