Refactor `cb login` to support browser login.
This is a major change to how the login flow works as well as how API
requests are authenticated.

Previously, `cb login` would prompt the user for a `client_id` and a
`client_secret`. Later it made the `client_id` optional given changes to
the format of the API key. The command would persist this key/secret and
then use it to procure access tokens when it was ready to start making
requests. Unfortunately this wasn't ideal as it 1) wasn't clear if a
login was successful and 2) logging out was confusing given the
relationship of the API key to access tokens. In summary, while it
served it's purpose this flow was no ideal in the long run. So we're
making it better.

Now, when running `cb login` the user will be prompted to open a browser
by pressing 'any key'.  Or they can abort the process by pressing `q`.
This is a pretty standard prompt so should be familiar to most. The
browser will open to a CLI login specific page at crunchybridge.com.
Here they will enter their credentials in the same way they would if
they were logging into the dashboard. Upon successful login, the user
can return to the CLI. Once the CLI has received notification that the
login attempt was successful, it will notify the user.

The process should result in something similar to the below example:

```
❯ cb login
Press any key to open a browser to login or q to exit:
Waiting for login... done
Logged in as user@example.com
```

Since it's likely that some users of `cb` require headless or scripted
interactions. We've added support for `CB_API_KEY` env. If this is
present then it will take precedence over any other credential that
exists on the system. This value is NOT stored by `cb` and so it will be
on the user to set the env appropriately or passing it along with calls.

For example:

```
❯ CB_API_KEY=cbkey_<secret> cb list
```

It is important to note the `cb login` command will not work if this env
is set. In fact, it will error accordingly.

```
❯ CB_API_KEY=cbkey_secret cb login
error: Cannot login with CB_API_KEY set.
```

Also as a result of these changes `cb token` will only work if an API
key is present.  This is not necessarily a restriction of `cb` but the
API itself as it will not issue access tokens unless the request to do
so is do so with an API key. This might change in the future, but for
now it is something that `cb` has to account for. Similar to the above
with `cb login`, `cb token` will notify the user appopriately if
attempting to procure an access token by any other means.  Generally, it
seems that this restriction for now is fine as most use of this command
are likely for scripting purposes only and an API key is likely already
in use.

This is just the first step in improving this flow. Followup efforts
will include enabling `keyring` support that will store the credential
in system supported security vaults. `Keychain` on MacOS and `GNOME
Keyring` on Linux. If neither of these two systems are supported on the
target OS, then we'll default back to the current `XDG_CONFIG_HOME`
approach. Though it's likely this will be rare at best.