Specify `order_field` parameter in cluster and team lists
The cluster and team list endpoints are exceptional in that unlike every
other endpoint, they order on `name` by default instead of `id`. This
is probably fine, but it'd be sort of nice to try and get all list
endpoints behaving the same way by default, so medium term I'm kind of
hoping I can change the default on cluster and team.

Reading cb code I wasn't totally able to determine whether it relies on
the default name ordering or not, but here, futureproof anyway by just
making sure to send an explicit `order_field=name` to these two
endpoints so they're always guaranteed to get today's ordering back.