Add login url to `cb login` output.
This came as a request in #139. The gist of it being that there are some
use cases where a browser may not be available to the user, e.g.
headless or no default browser. Therefore it would be beneficial to
present such configurations with a login url instead that can be
copy/pasted to a browser of choice.

Though, here, we're taking it one step further an making sure to provide
the link regardless of browser availability. Though, in the case of no
browser availability, we no longer prompt the user to open one and
instead just present the login url.

Example:

```
> # With browser
> cb login
Press Enter to open a browser to login. (Ctrl+C to quit)
Or visit: <login_url>

> # Without browser
> cb login
To login with Crunchy Bridge, please visit: <login_url>
```