Update `cb create` to require `--region` with `gcp` networks.
When creating a cluster in a specific network, there are a few rules
that apply that differ between the supported providers.

The execption is specifically with `gcp` based networks. GCP networks
are `global` meaning they are not bound to a specific `region`.
Therefore it is necessary to specify a region when creating a new
cluster of any type with a target network that is `gcp` based.

All other cases (`aws` and `azr`) a provider and region must not be
given along with a target network. As this information is attached/bound
to the network.