Changed `cb uri` and `cb psql` to default to `user` role.
Previously, the default for these commands when `--role` was not
specified was `default`. This role name directly maps to `postgres` and
is going to be deprecated and ultimately removed from the API as the
redundancy is confusing and unnecessary. So here, we are removing it as
an option as well as reducing the default role privilege to the lowest
level. If elevated privileges are necessary then they need to be
explicitly requested with the `--role` flag on each.