Update github release runner to include macos arm64
Previously, the github `macos-latest` runner was still an amd64/x86_64
architecture. As such we were not able to have a purely Apple Silicon
build through the github release workflow. However, such a runner now
exists that supports Apple Silicon[1][2]. However, the `-latest` tag
still points to the older images until sometime in Q2 2024[3]. So, we're
updating our workflows to be explicit in the `macos-<version>` to ensure
the correct architecture is utilized going forward. Also, we are
creating a create a specific build target to support Apple Silicon
builds using the `macos-14` runner in `.github/workflows/release.yml` as
well as updating the `build_macos_amd64` target to use the `macos-13`
runner. This will add an extra artifact to the releases which should
allow for an easier and perhaps more seamless homebrew release process.

[1] https://docs.github.com/en/actions/using-github-hosted-runners/about-github-hosted-runners/about-github-hosted-runners#standard-github-hosted-runners-for-public-repositories
[2] https://github.com/actions/runner-images#available-images
[3] https://github.blog/changelog/2024-01-30-github-actions-macos-14-sonoma-is-now-available/