Fix `cb psql` line wrapping issue.
It was reported in #158 that `cb psql` was having an issue with line
wrapping when text was entered beyond the bounds of the terminal window.
It turns out that this issue was related to characters in the prompt
being `Colorized`. And it was these bytes of the resulting string that
were causing the issue. So, we've simply removed the `Colorize` step
associated with this particular output. This does not impact the
coloring of the text in the prompt as that's part of the configured
`PROMPT1` in the resulting `.psqlrc`.

We've also take the opportunity here to update the uri fetching for the
`user` role which was missed as part of #162.