Add firewall rule management to `cb network` command.
These changes move firewall rule management under `cb network` and
deprecate the `cb firewall` rule command. This comes as part of a
request to be able to update firewall rules more completely as described
in #161.

The following commands are now supported:

* `cb network add-firewall-rule`
* `cb network list-firewall-rules`
* `cb network remove-firewall-rule`
* `cb network update-firewall-rule`

Each command requires the specification of the network to peform the
operation. As well, the shell completion logic has been updated to
suggest the list of available networks and firewall rule when
applicable.