Add `pagination` support for specific commands.
Add pagination support for specific commands to allow for list ALL the
resources they are associated with. They do NOT allow for the end user
to specify a limit/page size at this time. This is meant as an interim
fix to allow for users with large numbers of clusters, team, team
members and networks to be able list them all at one time.